+++ lib/factories.dart
@@ -3,7 +3,7 @@ import 'package:mock_cloud_firestore/mock_types.dart';
 import 'package:mockito/mockito.dart';
 
 MockQuerySnapshot createMockQuerySnapshot(Map<String, dynamic> colData,
-    {List<Map<String, dynamic>> added = const []}) {
+    {List<Map<String, dynamic>> added = const [], modified = const [], removed = const []}) {
   MockQuerySnapshot s = MockQuerySnapshot();
   List<MockDocumentChange> docChangeList = [];
   List<MockDocumentSnapshot> docSnapList = [];
@@ -13,7 +13,12 @@ MockQuerySnapshot createMockQuerySnapshot(Map<String, dynamic> colData,
   });
   added.forEach((value) {
     MockDocumentChange dc =
-        createDocumentChange(value, DocumentChangeType.added);
+    createDocumentChange(value, DocumentChangeType.added);
+    docChangeList.add(dc);
+  });
+  modified.forEach((value) {
+    MockDocumentChange dc =
+    createDocumentChange(value, DocumentChangeType.modified);
     docChangeList.add(dc);
   });
   when(s.documentChanges).thenAnswer((_) => docChangeList);